refactor: types file for common types
