+++ basket-stream/src/components/teams/TeamCard.tsx
@@ -1,16 +1,8 @@
+import { TeamType } from "@/lib/types";
 import { Box, Typography } from "@mui/material";
 import Image from "next/image";
+import Link from "next/link";
 import React from "react";
-
-export type TeamType = {
-  id: number;
-  name: string;
-  abbreviation: string;
-  conference: string;
-  division: string;
-  full_name: string;
-  imageSrc: string;
-};
 
 interface Props {
   team: TeamType;
@@ -18,48 +10,50 @@ interface Props {
 
 export default function TeamCard({ team }: Props) {
   return (
-    <Box
-      sx={{
-        width: "100%",
-        height: "100%",
-        backgroundColor: "common.black",
-
-        color: "common.white",
-        cursor: "pointer",
-        "&:hover": { backgroundColor: "primary.main", color: "common.black" },
-      }}
-    >
+    <Link href={`/teams/${team.abbreviation}`} style={{ textDecoration: "none" }}>
       <Box
         sx={{
-          position: "relative",
           width: "100%",
           height: "100%",
-          maxHeight: "90px",
-          p: "10px",
+          backgroundColor: "common.black",
+
+          color: "common.white",
+          cursor: "pointer",
+          "&:hover": { backgroundColor: "primary.main", color: "common.black" },
         }}
       >
-        <Image src={team.imageSrc} alt={team.name} fill style={{ objectFit: "contain" }} />
-      </Box>
-      <Box
-        sx={{
-          width: "100%",
-          height: { md: "50px", xs: "30px" },
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-          p: "0 10px",
-        }}
-      >
-        <Typography
+        <Box
           sx={{
-            textAlign: "center",
-            fontSize: { md: "14px", xs: "12px" },
-            fontWeight: 500,
+            position: "relative",
+            width: "100%",
+            height: "100%",
+            maxHeight: "90px",
+            p: "10px",
           }}
         >
-          {team.name}
-        </Typography>
+          <Image src={team.imageSrc} alt={team.name} fill style={{ objectFit: "contain" }} />
+        </Box>
+        <Box
+          sx={{
+            width: "100%",
+            height: { md: "50px", xs: "30px" },
+            display: "flex",
+            alignItems: "center",
+            justifyContent: "center",
+            p: "0 10px",
+          }}
+        >
+          <Typography
+            sx={{
+              textAlign: "center",
+              fontSize: { md: "14px", xs: "12px" },
+              fontWeight: 500,
+            }}
+          >
+            {team.name}
+          </Typography>
+        </Box>
       </Box>
-    </Box>
+    </Link>
   );
 }
