+++ basket-stream/src/lib/types.ts
@@ -1,0 +1,9 @@
+export type TeamType = {
+  id: number;
+  name: string;
+  abbreviation: string;
+  conference: string;
+  division: string;
+  full_name: string;
+  imageSrc: string;
+};
